feat(lms): add persistence trait
