+++ equed-lms/Classes/Domain/Model/Traits/PersistenceTrait.php
@@ -1,0 +1,65 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Equed\EquedLms\Domain\Model\Traits;
+
+use DateTimeImmutable;
+use Equed\Core\Service\ClockInterface;
+use Equed\Core\Service\UuidGeneratorInterface;
+use TYPO3\CMS\Extbase\Annotation\Inject;
+
+trait PersistenceTrait
+{
+    protected string $uuid = '';
+
+    protected DateTimeImmutable $createdAt;
+
+    protected DateTimeImmutable $updatedAt;
+
+    #[Inject]
+    protected UuidGeneratorInterface $uuidGenerator;
+
+    #[Inject]
+    protected ClockInterface $clock;
+
+    protected function initializePersistenceTrait(): void
+    {
+        if ($this->uuid === '') {
+            $this->uuid = $this->uuidGenerator->generate();
+        }
+        $now = $this->clock->now();
+        if (!isset($this->createdAt)) {
+            $this->createdAt = $now;
+        }
+        if (!isset($this->updatedAt)) {
+            $this->updatedAt = $now;
+        }
+    }
+
+    public function getUuid(): string
+    {
+        return $this->uuid;
+    }
+
+    public function getCreatedAt(): DateTimeImmutable
+    {
+        return $this->createdAt;
+    }
+
+    public function setCreatedAt(DateTimeImmutable $createdAt): void
+    {
+        $this->createdAt = $createdAt;
+    }
+
+    public function getUpdatedAt(): DateTimeImmutable
+    {
+        return $this->updatedAt;
+    }
+
+    public function setUpdatedAt(DateTimeImmutable $updatedAt): void
+    {
+        $this->updatedAt = $updatedAt;
+    }
+}
+
